Refactor: Move responseCodes to new constants.js file
