+++ utils/sendResponse.js
@@ -1,16 +1,8 @@
-const responseCodes = {
-  SUCCESS: 200,
-  CREATED: 201,
-  BAD_REQUEST: 400,
-  UNAUTHORIZED: 401,
-  FORBIDDEN: 403,
-  NOT_FOUND: 404,
-  INTERNAL_SERVER_ERROR: 500,
-};
+import { responseCodes } from "./constants.js";
 
 const sendResponse = (res, code, data) => {
   const statusCode = responseCodes[code];
   return res.status(statusCode).json(data);
 };
 
-export { responseCodes, sendResponse };
+export default sendResponse;
